+++ app/build.gradle
@@ -11,7 +11,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "org.befour.sms"
+        applicationId "org.positive.sms"
         minSdkVersion 23
         targetSdkVersion 30
         versionCode 1
