+++ app/src/main/AndroidManifest.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-	package="org.befour.sms">
+	package="org.positive.sms">
 	
 	<uses-permission android:name="android.permission.INTERNET" />
 	
 	<application
-		android:name=".SmsApplication"
+		android:name="org.positive.sms.SmsApplication"
 		android:allowBackup="true"
 		android:icon="@mipmap/ic_launcher"
 		android:label="@string/app_name"
 		android:roundIcon="@mipmap/ic_launcher_round"
 		android:supportsRtl="true"
 		android:theme="@style/Theme.SmsService">
-		<activity android:name=".presentation.main.MainActivity">
+		<activity android:name="org.positive.sms.presentation.main.MainActivity">
 			<intent-filter>
 				<action android:name="android.intent.action.MAIN" />
