--- app/src/main/java/org/befour/sms/SmsApplication.kt
@@ -1,7 +1,0 @@
-package org.befour.sms
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class SmsApplication : Application()
