--- app/src/main/java/org/befour/sms/data/model/ServerTimeResponse.kt
@@ -1,7 +1,0 @@
-package org.befour.sms.data.model
-
-import com.google.gson.annotations.SerializedName
-
-data class ServerTimeResponse(
-    @SerializedName("unixtime") val unixTime: Long
-)
