--- app/src/main/java/org/befour/sms/data/pref/AppSharedPreference.kt
@@ -1,6 +1,0 @@
-package org.befour.sms.data.pref
-
-interface AppSharedPreference {
-
-    var unixTime: Long
-}
