--- app/src/main/java/org/befour/sms/data/repository/ServerTimeRepository.kt
@@ -1,9 +1,0 @@
-package org.befour.sms.data.repository
-
-import io.reactivex.rxjava3.core.Single
-import org.befour.sms.data.model.ServerTimeResponse
-
-interface ServerTimeRepository {
-
-    fun getSeoulTime(): Single<ServerTimeResponse>
-}
