--- app/src/main/java/org/befour/sms/data/repository/ServerTimeRepositoryImpl.kt
@@ -1,11 +1,0 @@
-package org.befour.sms.data.repository
-
-import org.befour.sms.data.api.ServerTimeApi
-import javax.inject.Inject
-
-class ServerTimeRepositoryImpl @Inject constructor(
-    private val serverTimeApi: ServerTimeApi
-) : ServerTimeRepository {
-
-    override fun getSeoulTime() = serverTimeApi.loadServerTime("Asia/Seoul")
-}
