--- app/src/main/java/org/befour/sms/di/ContextModule.kt
@@ -1,16 +1,0 @@
-package org.befour.sms.di
-
-import android.app.Application
-import android.content.Context
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-object ContextModule {
-
-    @Provides
-    fun provideContext(application: Application): Context = application.applicationContext
-}
