--- app/src/main/java/org/befour/sms/extension/ActivityExtension.kt
@@ -1,8 +1,0 @@
-package org.befour.sms.extension
-
-import androidx.annotation.MainThread
-import org.befour.sms.presentation.base.BaseViewModel
-import org.befour.sms.presentation.base.BaseActivity
-
-@MainThread
-inline fun <reified VM : BaseViewModel> BaseActivity<*>.viewModelOf() = getViewModelLazy(VM::class)
