+++ app/src/main/java/org/positive/sms/SmsApplication.kt
@@ -1,0 +1,7 @@
+package org.positive.sms
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class SmsApplication : Application()
