+++ app/src/main/java/org/positive/sms/common/SingleLiveEvent.kt
@@ -1,0 +1,35 @@
+package org.positive.sms.common
+
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.Observer
+import java.util.concurrent.atomic.AtomicBoolean
+
+
+class SingleLiveEvent<T> : MutableLiveData<T>() {
+
+    private val pendingValue = AtomicBoolean(false)
+
+    override fun setValue(value: T) {
+        pendingValue.set(true)
+        super.setValue(value)
+    }
+
+    override fun postValue(value: T) {
+        pendingValue.set(true)
+        super.setValue(value)
+    }
+
+    override fun observe(owner: LifecycleOwner, observer: Observer<in T?>) {
+        super.observe(owner, { t ->
+            if (pendingValue.compareAndSet(true, false)) {
+                observer.onChanged(t)
+            }
+        })
+    }
+
+    fun call() {
+        pendingValue.set(true)
+        super.setValue(null)
+    }
+}
