+++ app/src/main/java/org/positive/sms/data/model/ServerTimeResponse.kt
@@ -1,0 +1,7 @@
+package org.positive.sms.data.model
+
+import com.google.gson.annotations.SerializedName
+
+data class ServerTimeResponse(
+    @SerializedName("unixtime") val unixTime: Long
+)
