+++ app/src/main/java/org/positive/sms/data/pref/AppSharedPreference.kt
@@ -1,0 +1,6 @@
+package org.positive.sms.data.pref
+
+interface AppSharedPreference {
+
+    var unixTime: Long
+}
