+++ app/src/main/java/org/positive/sms/di/ContextModule.kt
@@ -1,0 +1,16 @@
+package org.positive.sms.di
+
+import android.app.Application
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+object ContextModule {
+
+    @Provides
+    fun provideContext(application: Application): Context = application.applicationContext
+}
