+++ app/src/main/java/org/positive/sms/extension/ActivityExtension.kt
@@ -1,0 +1,8 @@
+package org.positive.sms.extension
+
+import androidx.annotation.MainThread
+import org.positive.sms.presentation.base.BaseViewModel
+import org.positive.sms.presentation.base.BaseActivity
+
+@MainThread
+inline fun <reified VM : BaseViewModel> BaseActivity<*>.viewModelOf() = getViewModelLazy(VM::class)
