+++ app/src/main/java/org/positive/sms/presentation/main/MainActivity.kt
@@ -1,0 +1,19 @@
+package org.positive.sms.presentation.main
+
+import android.os.Bundle
+import dagger.hilt.android.AndroidEntryPoint
+import org.positive.sms.R
+import org.positive.sms.databinding.ActivityMainBinding
+import org.positive.sms.extension.viewModelOf
+import org.positive.sms.presentation.base.BaseActivity
+
+@AndroidEntryPoint
+class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
+
+    private val viewModel by viewModelOf<MainViewModel>()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding.viewModel = viewModel
+    }
+}
