+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
 		
 		<variable
 			name="viewModel"
-			type="org.befour.sms.presentation.main.MainViewModel" />
+			type="org.positive.sms.presentation.main.MainViewModel" />
 	</data>
 	
 	<androidx.constraintlayout.widget.ConstraintLayout
